% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{et_array}
\alias{et_array}
\title{Declare a tracked array for \code{\link[=et_aggregate]{et_aggregate()}}.}
\usage{
et_array(type = c("Int", "Float64", "Bool"), dim)
}
\arguments{
\item{type}{Julia element type: \code{"Int"}, \code{"Float64"} or \code{"Bool"}.}

\item{dim}{Integer vector of dimensions.}
}
\value{
An object of class \code{et_array}.
}
\description{
Declare a tracked array for \code{\link[=et_aggregate]{et_aggregate()}}.
}
