% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfo.R
\name{et_lfo_elpd}
\alias{et_lfo_elpd}
\title{Total ELPD from a leave-future-out sweep.}
\usage{
et_lfo_elpd(res, granularity = res$granularities[1])
}
\arguments{
\item{res}{An \code{\link[=et_lfo_cv]{et_lfo_cv()}} result.}

\item{granularity}{Which granularity to total; defaults to the first one run.}
}
\value{
A single number.
}
\description{
Total ELPD from a leave-future-out sweep.
}
