% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{et_nuts}
\alias{et_nuts}
\title{A NUTS block.}
\usage{
et_nuts(vars, target_accept = 0.8)
}
\arguments{
\item{vars}{Character vector of parameter names sampled together.}

\item{target_accept}{Target acceptance rate.}
}
\value{
An \code{et_block}.
}
\description{
A NUTS block.
}
