% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpile_fn.R
\name{et_protocols}
\alias{et_protocols}
\title{The call protocols a user function can be transpiled into.}
\usage{
et_protocols()
}
\value{
A character vector of role names.
}
\description{
The call protocols a user function can be transpiled into.
}
