% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{et_trans}
\alias{et_trans}
\title{Declare one transition.}
\usage{
et_trans(from, to, rate)
}
\arguments{
\item{from, to}{State names.}

\item{rate}{A rate function \verb{(model, data, i, t)} or an \code{\link[=et_julia]{et_julia()}}.}
}
\value{
An object of class \code{et_trans}.
}
\description{
Declare one transition.
}
