% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpile.R
\name{et_transpile_expr}
\alias{et_transpile_expr}
\title{Transpile one R expression (language object) to Julia source.}
\usage{
et_transpile_expr(e, ctx = new_ctx())
}
\arguments{
\item{e}{An R language object, symbol or literal, e.g. from \code{body(f)}.}

\item{ctx}{A context from \code{\link[=new_ctx]{new_ctx()}}.}
}
\value{
A length-1 character string of Julia source.
}
\description{
Transpile one R expression (language object) to Julia source.
}
