% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_distributions}
\alias{prior_distributions}
\alias{normal_dist}
\alias{gamma_dist}
\alias{beta_dist}
\alias{exponential_dist}
\alias{uniform_dist}
\alias{dirichlet_dist}
\alias{truncated_dist}
\alias{custom_dist}
\alias{et_normal}
\alias{et_gamma}
\alias{et_beta}
\alias{et_exponential}
\alias{et_uniform}
\alias{et_dirichlet}
\alias{et_truncated}
\alias{et_dist}
\title{Prior distributions.}
\usage{
normal_dist(mean = 0, sd = 1)

gamma_dist(shape, scale)

beta_dist(a = 1, b = 1)

exponential_dist(rate = 1)

uniform_dist(a = 0, b = 1)

dirichlet_dist(alpha)

truncated_dist(d, lower = NA, upper = NA)

custom_dist(name, args = list())

et_normal(...)

et_gamma(...)

et_beta(...)

et_exponential(...)

et_uniform(...)

et_dirichlet(...)

et_truncated(...)

et_dist(...)
}
\arguments{
\item{mean, sd}{Normal parameters.}

\item{shape, scale}{Gamma parameters (Julia's \code{Gamma(shape, scale)}).}

\item{a, b}{Beta / Uniform parameters.}

\item{rate}{Exponential rate; emitted as \code{Exponential(1 / rate)} because
Julia's \code{Exponential} is parameterised by scale, which is the single most
likely thing for an R user to get backwards.}

\item{alpha}{Dirichlet concentration vector.}

\item{d}{A distribution to truncate.}

\item{lower, upper}{Truncation bounds; \code{NA} for unbounded on that side.}

\item{name, args}{Escape hatch: a Julia distribution name and its arguments.}
}
\value{
An object of class \code{custom_dist}.
}
\description{
These map onto \code{Distributions.jl} constructors of the same name.
}
