% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{et_julia_project}
\alias{et_julia_project}
\title{The pinned Julia project shipped with this package.}
\usage{
et_julia_project()
}
\value{
A path.
}
\description{
The pinned Julia project shipped with this package.
}
