% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{et_priors}
\alias{et_priors}
\alias{et_normal}
\alias{et_gamma}
\alias{et_beta}
\alias{et_exponential}
\alias{et_uniform}
\alias{et_dirichlet}
\alias{et_truncated}
\alias{et_dist}
\title{Prior distributions.}
\usage{
et_normal(mean = 0, sd = 1)

et_gamma(shape, scale)

et_beta(a = 1, b = 1)

et_exponential(rate = 1)

et_uniform(a = 0, b = 1)

et_dirichlet(alpha)

et_truncated(d, lower = NA, upper = NA)

et_dist(name, args = list())
}
\arguments{
\item{mean, sd}{Normal parameters.}

\item{shape, scale}{Gamma parameters (Julia's \code{Gamma(shape, scale)}).}

\item{a, b}{Beta / Uniform parameters.}

\item{rate}{Exponential rate; emitted as \code{Exponential(1 / rate)} because
Julia's \code{Exponential} is parameterised by SCALE, which is the single most
likely thing for an R user to get backwards.}

\item{alpha}{Dirichlet concentration vector.}

\item{d}{A distribution to truncate.}

\item{lower, upper}{Truncation bounds; \code{NA} for unbounded on that side.}

\item{name, args}{Escape hatch: a Julia distribution name and its arguments.}
}
\value{
An object of class \code{et_dist}.
}
\description{
These map onto \code{Distributions.jl} constructors of the same name.
}
